🔧 chore(treesitter.rc.lua): add support for Dart language in treesitter 🔧 chore(base.lua): comment out textwidth and colorcolumn options ✨ feat(lspconfig.lua): add support for Dart Language Server (dartls) 🔧 chore(lspconfig.lua): reformat code for better readability
The changes in `treesitter.rc.lua` add support for the Dart language in the treesitter plugin configuration.

In `base.lua`, the `textwidth` and `colorcolumn` options are commented out to disable them.

The changes in `lspconfig.lua` add support for the Dart Language Server (dartls) and reformat the code for better readability.